Remove (now outdated/incomplete) going from Google doc to markdown file
ATM most of the changes might be changes to the schema, not as
much to markdown text files. In the future we might acquire some other
features (e.g. workflow figures etc) which would also need to be changed.
Overall I think we should strive to remove as much of technical aspects
from governance principles as possible -- they should be outlined in other
documents, such as https://github.com/bids-standard/bids-extensions etc