style: use top-to-bottom orientation and hierarchical layout
This adjustment will allow items in the
mermaid diagram to appear bigger in the
webpage because the page is more limited
in the horizontal dimension. Additionally,
the syntax of the mermaid diagram has been
updated to the latest syntax,
e.g. `flowchart` instead of `graph`.
The intents of the lines in the diagram
are those recommended/defaulted by the
mermaid live editor at
https://www.mermaidchart.com