Change wording to point to not limit BIDS to "neuroimaging"
BIDS has expanded to cover various data types which are not strictly "imaging" but relate
to neural data or expressions of neural processes: e.g. purely behavioral datasets, ongong
work on BEP032 on microelectrodes arrays, etc.  As such, we should start moving away from
limiting description of BIDS only to neuroimaging.  This would be the first step in a sequence
of such changes and I think it does not need tobe  contingent on any other change